[#153] Move metadata+views to their own module
As mentioned in the last commit's message, here we're moving the
contract's metadata+views to their own module.
